requirements.txt
Flask==3.0.2
redis==5.2.1
rq==2.1.0
gunicorn==23.0.0
torch==2.4.0
torchaudio==2.4.0
torchvision==0.19.0
transformers==4.43.3
python-dotenv==1.0.1